requirements.txt
Flask==1.1.1
moviepy==1.0.1
Keras==2.2.4
torchvision==0.4.2
tensorflow_gpu==1.15.0
pandas==0.25.3
scipy==1.3.2
torchsummary==1.5.1
opencv_contrib_python==4.1.2.30
torch==1.3.1
numpy==1.17.4
Pillow==7.0.0
scikit_learn==0.22.2.post1
tensorflow==2.1.0
